+++ src/Attribute/Inject.php
@@ -3,6 +3,6 @@
 namespace Nandan108\DtoToolkit\Attribute;
 
 #[\Attribute(\Attribute::TARGET_PROPERTY)]
-final class Injected
+final class Inject
 {
 }